% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_item.R
\name{add_item}
\alias{add_item}
\title{Add an item to the pool of sensors, loggers, batteries}
\usage{
add_item(item, ID, UID, type, date, capacity, comment)
}
\arguments{
\item{item}{\code{Character} value, item to be added. Must be one of
\code{"sensor"}, \code{"logger"}, \code{"battery"}.}

\item{ID}{\code{Character} value, optional manufacturer ID or label of the
item to add. IDs will not be used by 'eseisManager' for object
identification but only be shown in reports.}

\item{UID}{\code{Character} value, mandatory unique identifier of the item
to add. UIDs will be used by 'eseisManager' to identify objects.}

\item{type}{\code{Character} value, mandatory information on the type of
sensor, logger or battery. Entry will be used in pull down menu of the
GUI. Hence, the types of items should be named consistently.}

\item{date}{\code{POSIXct} or \code{Character} value, date when
the item was added to the pool. Character strings will be converted to
POSIXct format if possible. It is sufficient to define year, month and
day.}

\item{capacity}{\code{Character} value, depending on the type of item,
this information can have different meanings. For \code{item = "sensor"}
capacity will be ignored. For \code{item = "logger"} it will imply the
logger's storage capacity (e.g. \code{"64 GB"}). For \code{item = "battery"}
it will imply the battery's capacity (e.g. \code{"200 Ah"}).}

\item{comment}{\code{Character} string, optional comment line which will
be stored in the system text file but not be tracked or used by the
package reports.}
}
\description{
The function adds an entry to the ASCII files which will be used
by the package for further workflows.
}
\examples{

\dontrun{

add_item(item = "battery",
         ID = "PowerCell",
         UID = "1999-01-01-08154712",
         type = "AGM",
         date = "1999-01-01",
         capacity = "55 Ah",
         comment = "dummy entry")

}

}
\author{
Michael Dietze
}
\keyword{eseis}
