% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_name.R
\name{add_name}
\alias{add_name}
\title{Add a researcher name to the data base}
\usage{
add_name(name, sort = TRUE)
}
\arguments{
\item{name}{\code{Character} value, given name and family name of the
person to add, e.g. \code{"Michael Dietze"}.}

\item{sort}{\code{Logical} value, option to update the order of names
in the list of researchers. Default is \code{TRUE}.}
}
\description{
The function adds a name of a potential responsible researcher, an
information needed for every new commit.
}
\examples{

\dontrun{

add_name(name = "Heinz Ketchup")

}

}
\author{
Michael Dietze
}
\keyword{eseis}
