% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_project.R
\name{add_project}
\alias{add_project}
\title{Add a project to the data base}
\usage{
add_project(name, title, lat, lon, start)
}
\arguments{
\item{name}{\code{Character} value, short name or acronym of the project
to add}

\item{title}{\code{Character} value, one line summary of the project,
ideally expressing where and what the project is about.}

\item{lat}{\code{Numeric} value, latitude of the project areas centroid}

\item{lon}{\code{Numeric} value, longitude of the project areas centroid}

\item{start}{\code{POSIXct} value, start time of the project. A character
string will be (attempted to be) converted to the POSIXct format. It is
sufficient to provide year, month and day.}
}
\description{
The function adds project with meta information to the data base system
text file.
}
\examples{

\dontrun{

add_project(name = "Earthworms",
            title = "Seismic earthworm tracking in Patagonia",
            lat = -45
            lon = -72,
            start = as.POSIXct("1921-04-01"))

}

}
\author{
Michael Dietze
}
\keyword{eseis}
