% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_validity.R
\name{check_validity}
\alias{check_validity}
\title{Check activity commits for validity}
\usage{
check_validity(commits)
}
\arguments{
\item{commits}{\code{Character} value, file name of the commit log file
to screen.}
}
\value{
\code{Character} vector with identified issues.
}
\description{
The function screens a commit log for potential validity violations. The
identified issues are returned.
}
\details{
The following tests are performed:
}
\examples{

\dontrun{
# Start the GUI
check_validity(commits = "/path/to/sourcefiles/commits.txt")
}

}
\author{
Michael Dietze
}
\keyword{eseis}
