% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backup.R
\name{create_backup}
\alias{create_backup}
\title{Create source files backup}
\usage{
create_backup(path_save, sourcefiles)
}
\arguments{
\item{path_save}{\code{Character} value, path where the backup archive
will be saved. By default, the data is saved in
\code{Documents/projects/Environmental_seismology/00_stationtracker/backup}}

\item{sourcefiles}{\code{Character} value, path to the directory that
contains the source files of the data base. If omitted, it is assumed
that the source files are present in the extdata directory of the R
package.}
}
\value{
A zip file with source files backup
}
\description{
The function creates a compressed backup version of the source
files.
}
\details{
It is recommended to run this function after new commits have been
added.
}
\examples{

\dontrun{
create_backup()
}

}
\author{
Michael Dietze
}
\keyword{eseis}
