% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseisManager-package.R
\docType{package}
\name{eseisManager}
\alias{eseisManager}
\alias{eseisManager-package}
\title{eseisManager: Seismic Station Pool Management Toolbox}
\description{
The package provides functions to maintain a
database of activities of seismic stations.
}
\details{
\tabular{ll}{
**Package:** \tab eseisManager \cr
**Type:** \tab Package \cr
**Version:** \tab 0.1.0 \cr
**Date:** \tab 2020-08-28 \cr
**License:** \tab GPL-3 \cr
}
}
\author{
Michael Dietze
}
\keyword{package}
