% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_input.R
\name{gui_input}
\alias{gui_input}
\title{User interface to add new activities to the pool lifeline}
\usage{
gui_input()
}
\description{
The function opens a graphical user interface to conveniently update
changes to the seismic pool items. The function must be called from
the directory which contains the management source files (see details).
}
\details{
The function requires a set of predefined source files. These source
files must be located in the same directory. The following files are
required:

\itemize{
  \item \code{commits.txt}, ASCII file, tab-separated, containing at
  least two lines of (dummy) data. The following column names must be
  present: \code{ID}, \code{date}, \code{name}, \code{project},
  \code{date_start}, \code{date_stop}, \code{type_station},
  \code{id_station}, \code{id_sensor}, \code{id_logger}, \code{type_sensor},
  \code{type_logger}, \code{BOB}, \code{type_battery}.
  \item \code{projects.txt}, ASCII file, comma separated, containing
  information about projects to choose from. The following column names
  must be present: \code{name}, \code{running_title}, \code{lat},
  \code{lon}.
  \item \code{names.txt}, ASCII file with scientist names to choose from,
  one per line.
  \item \code{sensors.txt},
  \item \code{loggers.txt},
  \item \code{batteries.txt},
}
}
\examples{

\dontrun{
# Start the GUI
gui_input()
}

}
\author{
Michael Dietze
}
\keyword{eseis}
