% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_item.R
\name{report_item}
\alias{report_item}
\title{Generate a report on a selected item of the seismic pool}
\usage{
report_item(
  item = "station",
  UID,
  filename,
  browser = TRUE,
  extent,
  sourcefiles,
  interactive = TRUE
)
}
\arguments{
\item{item}{\code{Character} value, item about which to report, one
out of \code{"station"}, \code{"sensor"}, \code{"logger"}.}

\item{UID}{\code{Character} value, UID of the item to report on (i.e. ID
of a station and UID of a logger or sensor).}

\item{filename}{\code{Character} value, name of the output file (without
extension). Default is \code{"report"}, saved in a temporary directory.
Please note that no ~ signs are allowed as path flags.}

\item{browser}{\code{Logical} value, option to open report in browser.
Default is \code{TRUE}.}

\item{extent}{\code{Numeric} vector, extent of the output map. Must be
in the order xmin, xmax, ymin, ymax. If omitted, global map will be
used.}

\item{sourcefiles}{\code{Character} value, path to the directory that
contains the source files of the data base. See details for further
information.}

\item{interactive}{\code{Logical} value, option to use an interactive map
instead of a static world map. Default is \code{TRUE}.}
}
\value{
HTML document and data frame with requested information.
}
\description{
By default, the source files are stored in the extdata
directory of the installed R package.
}
\examples{

\dontrun{
report_item(item = "station", ID = "HID01")
}

}
\author{
Michael Dietze
}
\keyword{eseis}
